review2/requirements.txt
# Review II Scraper Dependencies

# Core scraping libraries
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# Configuration and data handling
PyYAML>=6.0
pandas>=1.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Optional: Advanced features
selenium>=4.0.0  # For JavaScript-heavy sites
aiohttp>=3.8.0   # For async scraping (future enhancement)

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991